Added thread safe cache

+++ HackerNews.Api/Services/HackerNewsClient.cs
@@ -6,23 +6,23 @@ namespace HackerNews.Api.Services;
 
 public interface IHackerNewsClient
 {
-    Task<IEnumerable<int>> GetBestIds(CancellationToken cancellationToken);
-    IAsyncEnumerable<Story> GetStories(HashSet<int> ids, CancellationToken cancellationToken);
-    Task<Story> GetStory(int id, CancellationToken cancellationToken);
+    Task<IEnumerable<long>> GetBestIds(CancellationToken cancellationToken);
+    IAsyncEnumerable<Story> GetStories(HashSet<long> ids, CancellationToken cancellationToken);
+    Task<Story> GetStory(long id, CancellationToken cancellationToken);
 }
 
 public class HackerNewsClient : IHackerNewsClient
 {
     private readonly HttpClient _httpClient;
     private const string GetBestStoriesUrl = "beststories.json";
-    private string GetStoryUrl(int id) => $"item/{id}.json";
+    private string GetStoryUrl(long id) => $"item/{id}.json";
 
     public HackerNewsClient(HttpClient httpClient)
     {
         _httpClient = httpClient;
     }
     
-    public async Task<IEnumerable<int>> GetBestIds(CancellationToken cancellationToken)
+    public async Task<IEnumerable<long>> GetBestIds(CancellationToken cancellationToken)
     {
         var response = await _httpClient.GetAsync(GetBestStoriesUrl, cancellationToken);
 
@@ -32,7 +32,7 @@ public class HackerNewsClient : IHackerNewsClient
         }
         
         var jsonString = await response.Content.ReadAsStringAsync(cancellationToken);
-        var stories = JsonConvert.DeserializeObject<List<int>>(jsonString);
+        var stories = JsonConvert.DeserializeObject<List<long>>(jsonString);
         if (stories is null)
         {
             throw new Exception("Could not get stories");
@@ -41,7 +41,7 @@ public class HackerNewsClient : IHackerNewsClient
         return stories;
     }
 
-    public async IAsyncEnumerable<Story> GetStories(HashSet<int> ids, [EnumeratorCancellation] CancellationToken cancellationToken)
+    public async IAsyncEnumerable<Story> GetStories(HashSet<long> ids, [EnumeratorCancellation] CancellationToken cancellationToken)
     {
         var tasks = ids.Select(id => GetStory(id, cancellationToken));
 
@@ -53,7 +53,7 @@ public class HackerNewsClient : IHackerNewsClient
         }
     }
 
-    public async Task<Story> GetStory(int id, CancellationToken cancellationToken)
+    public async Task<Story> GetStory(long id, CancellationToken cancellationToken)
     {
         var response = await _httpClient.GetAsync(GetStoryUrl(id), cancellationToken);
